Add start page for labs
This adds a new page, linked to from the home page, where users can read instructions for how to hire labs and download the CSV file of user labs.
