+++ app/templates/index.html
@@ -34,12 +34,12 @@
             "body": "eg a booking system or accessibility audit",
           },
           {
-            "link": url_for("buyers.info_page_for_starting_a_brief", framework_slug='digital-outcomes-and-specialists', lot_slug='user-research-participants'),
+            "link": url_for("main.info_page_for_starting_a_brief", framework_slug='digital-outcomes-and-specialists', lot_slug='user-research-participants'),
             "title": "Find user research participants",
             "body": "eg people from a specific user group to test your service",
           },
           {
-            "link": url_for("main.info_page_for_starting_a_brief", framework_slug='digital-outcomes-and-specialists', lot_slug='user-research-studios'),
+            "link": url_for("main.studios_start_page", framework_slug='digital-outcomes-and-specialists'),
             "title": "Find a user research lab",
             "body": "eg a room to conduct research sessions",
           },
